==> 16. Text Classification Project/Step 2. Build Dictionary.ipynb <==
from pprint import pprint
from nltk.corpus import words as nltk_words

english_dictionary = {word : True for word in list(set(nltk_words.words()))}

def is_in_english(word):
    try:
        return english_dictionary[word]
    except:
        return False

"""
Build Dictionary of Words
"""
document_paths = []
from pprint import pprint
import os
walk = os.walk('.\\clean_data\\20_newsgroups', topdown = False)
for root, dirs, files in walk :
    for file in files:
        doc = {}
        dir_ = root.split('\\')[-1]
        doc['path'] = root+"\\"+file
        doc['target'] = dir_
        document_paths.append(doc)
pprint(document_paths[0:10000:100])


import pandas as pd
df = pd.DataFrame(document_paths)
df.head()

targets = list(set(df['target'].values))
pprint(targets)
class_dict = { i : targets[i] for i in range(len(targets))}
pprint(class_dict)


class_df = pd.DataFrame()
class_df['class'] = class_dict
class_df['category'] = [class_dict[key] for key in class_dict]
class_df.to_csv("class_dict.csv", index = False)

df['class'] = [targets.index(target) for target in df['target'].values]

df.head()

X = df['path']
Y = df['class']
print(X[0:5])
print(Y[0:5])

from sklearn.model_selection import train_test_split
X_train, X_test, Y_train, Y_test = train_test_split(X, Y, random_state = 0) 
print(X_train.shape)
print(X_test.shape)
print(Y_train.shape)
print(Y_test.shape)


train_df = pd.DataFrame()
train_df['path'] = X_train
train_df['class'] = Y_train

train_df.to_csv("train_files.csv", index = False)

test_df = pd.DataFrame()
test_df['path'] = X_test
test_df['class'] = Y_test
test_df.to_csv("test_files.csv", index = False)


train_df.head()

train_df.head()

test_df.head()

master_dictionary = {}

i = 0
import time
st = time.time()
new_paths = []
for doc_path in X_train:
    #if i == 1 :
    #    break
    path = doc_path
    with open(path) as doc :
        num_tokens_in_doc = 0
        data = ''.join(doc.readlines()).split('.')
        for line in data:
            for word in line.split():
                if not is_in_english(word):
                    continue
                #print(word)
                try:
                    master_dictionary[word] += 1
                except KeyError:
                    master_dictionary[word] = 1
                    continue
        doc.close()
        i += 1
        if i%500 == 0 :
            print( i, "Files Processed in", int((time.time() - st)*1000) , "millisec")
            st = time.time()
#print(master_dictionary)

import numpy as np
keys = list(master_dictionary)
freq = np.array([ master_dictionary[key] for key in keys])

words_df = pd.DataFrame()
words_df['word'] = keys
words_df['frequency'] = freq
words_df.to_csv('all_words.csv', index = False)
