==> 16. Text Classification Project/Step 3. Build Dataset.ipynb <==
"""
Build Dataset
"""

import numpy as np
import pandas as pd
max_allowed_freq = 16321
min_allowed_freq = 20

class_dict_df = pd.DataFrame(pd.read_csv("class_dict.csv"))
class_dict_df.head()

#selected_words_df = pd.DataFrame(pd.read_csv('all_words.csv'))
selected_words_df = pd.DataFrame(pd.read_csv('selected_words_'+ str(max_allowed_freq) +'_'+ str(min_allowed_freq) +'.csv'))
selected_words_df.head()


selected_words = selected_words_df['word'].values
num_selected_words = len(selected_words)


train_file_paths_df = pd.DataFrame(pd.read_csv("train_files.csv"))
train_file_paths_df.head()

train_file_paths = train_file_paths_df['path'].values
train_file_class = train_file_paths_df['class'].values

import time
matrix = []
st = time.time()
for i in range(len(train_file_paths)) :
    #if i == 2000:
    #    break
    if i % 1000 == 0 :
        et = time.time()
        print( i, "Files Processed in", int((time.time() - st)*1000) , "millisec")
        st = time.time()    
    path = train_file_paths[i]
    target = train_file_class[i]
    with open(path) as file :
        X = []
        count = {}
        data = file.readlines()
        words_in_file = (''.join(data)).split(' ')
        for word in words_in_file:
            if word in selected_words:
                try:
                    count[word] += 1
                except:
                    count[word] = 1
        for word in selected_words:
            try :
                x = count[word] #X.append(count[word])
            except :
                count[word] = 0 #X.append(0)
        #print(X)
        X = [count[word] for word in selected_words if word in count] 
        #print(len(X))
        matrix.append(X)
        #X_train_df.iloc[i, :] = 
        file.close()
print(len(matrix))
print(len(matrix[0]))

X_train_df = pd.DataFrame(matrix, columns = selected_words)

X_train_df.describe()

import copy
X_train = X_train_df.values
Y_train = copy.deepcopy(train_file_class)
print(X_train.shape)
print(Y_train.shape)

dataset = pd.DataFrame(X_train_df)
dataset['target'] = Y_train
dataset.to_csv("20_newsgroups_dataset.csv", index = False)

# , fit_prior = False 0.815763152631
# 
from sklearn.naive_bayes import MultinomialNB
model = MultinomialNB(alpha = 1)
model.fit(X_train, Y_train)

print(model.score(X_train, Y_train))

from sklearn.metrics import confusion_matrix

Y_pred = model.predict(X_train)

from pprint import pprint
pprint(confusion_matrix(Y_pred, Y_train))

for target in set(Y_train):
    print(target, ":", (Y_train == target).sum())

from sklearn.metrics import classification_report
print(classification_report(Y_pred, Y_train))

import _pickle as pickle

model_data = {}
model_data['model'] = model
model_data['max_freq'] = max_allowed_freq
model_data['min_freq'] = min_allowed_freq
with open("multinomial_naive_bayes_" + str(max_allowed_freq) + "_" + str(min_allowed_freq) + ".sav", 'wb') as model_file :
    pickle.dump(model_data, model_file )
    model_file.close()
